Calculate with precisely functions across values

map_precisely() is a wrapper around tidyr::crossing() and purrr::pmap_dfr() to give a set of values to any of the calculation functions in precisely. All possible combinations of the values are passed to the function, returning a tibble where each row is the result for each combination.

map_precisely(.f, ...)

Arguments

.f
a function in precisely

...
arguments passed to .f. All possible combinations of argument values are given to the function.

Value

a tibble

Examples

map_precisely(
  n_risk_difference,
  precision = seq(from = .02, to = .20, by = .005),
  exposed = c(.2, .4),
  unexposed = c(.1, .3),
  group_ratio = 1
)

#> [90m# A tibble: 148 x 8[39m
#>    n_exposed n_unexposed n_total risk_difference precision exposed unexposed
#>        [3m[90m<dbl>[39m[23m       [3m[90m<dbl>[39m[23m   [3m[90m<dbl>[39m[23m           [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m   [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m
#> [90m 1[39m     [4m9[24m604.       [4m9[24m604.  [4m1[24m[4m9[24m207.           0.1       0.02      0.2       0.1
#> [90m 2[39m    [4m1[24m[4m4[24m213.      [4m1[24m[4m4[24m213.  [4m2[24m[4m8[24m427.          -[31m0[39m[31m.[39m[31m100[39m     0.02      0.2       0.3
#> [90m 3[39m    [4m1[24m[4m2[24m677.      [4m1[24m[4m2[24m677.  [4m2[24m[4m5[24m354.           0.3       0.02      0.4       0.1
#> [90m 4[39m    [4m1[24m[4m7[24m287.      [4m1[24m[4m7[24m287.  [4m3[24m[4m4[24m573.           0.1       0.02      0.4       0.3
#> [90m 5[39m     [4m6[24m146.       [4m6[24m146.  [4m1[24m[4m2[24m293.           0.1       0.025     0.2       0.1
#> [90m 6[39m     [4m9[24m097.       [4m9[24m097.  [4m1[24m[4m8[24m193.          -[31m0[39m[31m.[39m[31m100[39m     0.025     0.2       0.3
#> [90m 7[39m     [4m8[24m113.       [4m8[24m113.  [4m1[24m[4m6[24m226.           0.3       0.025     0.4       0.1
#> [90m 8[39m    [4m1[24m[4m1[24m063.      [4m1[24m[4m1[24m063.  [4m2[24m[4m2[24m127.           0.1       0.025     0.4       0.3
#> [90m 9[39m     [4m4[24m268.       [4m4[24m268.   [4m8[24m537.           0.1       0.03      0.2       0.1
#> [90m10[39m     [4m6[24m317.       [4m6[24m317.  [4m1[24m[4m2[24m634.          -[31m0[39m[31m.[39m[31m100[39m     0.03      0.2       0.3
#> [90m# ... with 138 more rows, and 1 more variable: group_ratio [3m[90m<dbl>[90m[23m[39m